+++ CHANGELOG.md
@@ -13,6 +13,7 @@
   * Testing all possible variations of scrap scale option parsing.
   * Custom exceptions should provide trace.
   * Renaming _execPrePrepareUndoRedoInfo and _execPreCreateUndoRedoCommand at MPCommand.
+  * Creating MPAddLineSegmentCommand.atExistingLineSegmentPosition().
 
 ## 0.2.35 - 2026-02-13 - The [Chiquita Bacana's daughter](https://www.youtube.com/watch?v=QajgvvuFOSg) release
 * New features: